First round of refactor
