+++ src/coguard_cli/image_check/__init__.py
@@ -11,9 +11,8 @@ import logging
 import tempfile
 from typing import Optional, Dict, Tuple
 import zipfile
-from coguard_cli.util import replace_special_chars_with_underscore, \
-    create_service_identifier
-from coguard_cli.auth.util import DealEnum
+from coguard_cli.check_common_util import replace_special_chars_with_underscore
+from coguard_cli.util import create_service_identifier
 from coguard_cli import docker_dao
 import coguard_cli.discovery.config_file_finder_factory as factory
 from coguard_cli.print_colors import COLOR_RED, COLOR_TERMINATION
