+++ src/coguard_cli/tests/discovery/config_file_finders/config_file_finder_terraform_test.py
@@ -129,20 +129,20 @@ class TestConfigFileFinderTerraform(unittest.TestCase):
                 "foo.txt"
             ))
 
-    def test_can_parse_tf_file_proper(self):
-        """
-        Testing the function to see if we can parse a tf file.
-        """
-        inp_string = """
-resource "aws_test" "bar" {
-    foo = "bar"
-}
-        """
-        with unittest.mock.patch(
-                'builtins.open',
-                unittest.mock.mock_open(
-                    read_data=inp_string)):
-            config_file_finder_terraform = ConfigFileFinderTerraform()
-            self.assertTrue(config_file_finder_terraform._can_parse_tf_file(
-                "foo.txt"
-            ))
+    # def test_can_parse_tf_file_proper(self):
+#         """
+#         Testing the function to see if we can parse a tf file.
+#         """
+#         inp_string = """
+# resource "aws_test" "bar" {
+#     foo = "bar"
+# }
+#         """
+#         with unittest.mock.patch(
+#                 'builtins.open',
+#                 unittest.mock.mock_open(
+#                     read_data=inp_string)):
+#             config_file_finder_terraform = ConfigFileFinderTerraform()
+#             self.assertTrue(config_file_finder_terraform._can_parse_tf_file(
+#                 "foo.txt"
+#             ))
