+++ src/coguard_cli/util.py
@@ -3,34 +3,31 @@ Common utilities throughout the project.
 """
 
 import os
+import sys
 import json
 import logging
-import re
 import pathlib
 from enum import Enum
+import textwrap
 
 import shutil
-from typing import Set, Dict, Optional, Tuple
+from typing import Set, Dict, Optional, Tuple, List
+from coguard_cli.print_colors import COLOR_TERMINATION, \
+    COLOR_RED, COLOR_CYAN, COLOR_YELLOW, COLOR_GRAY
+from coguard_cli import api_connection
+from coguard_cli.output_generators.output_generator_sarif import \
+    translate_result_to_sarif
+from coguard_cli.output_generators.output_generator_markdown import \
+    translate_result_to_markdown
+from coguard_cli.auth.token import Token
+from coguard_cli.auth.enums import DealEnum
+from coguard_cli.auth.auth_config import CoGuardCliConfig
 
 class CiCdProviderNames(Enum):
     """
     The enumeration capturing the different CI/CD Provider Names.
     """
     GITHUB = "github"
-
-def replace_special_chars_with_underscore(string: str, keep_spaces=False) -> str:
-    """
-    Helper function remove any special character with underscore.
-    """
-    return re.sub(
-        "_+",
-        "_",
-        re.sub(
-            "[^a-zA-Z1-9- ]" if keep_spaces else "[^a-zA-Z1-9]",
-            "_",
-            string
-        )
-    ).strip("_ ")
 
 def create_service_identifier(prefix: str,
                               currently_used_service_names: Set[str],
@@ -168,3 +165,217 @@ def convert_posix_path_to_os_path(input_str: str) -> str:
     current path separator is.
     """
     return os.sep.join(pathlib.PurePosixPath(input_str).parts)
+
+def retrieve_coguard_ignore_values(
+        folder_name: str) -> List[str]:
+    """
+    Helper function to return the coguard_ignore_values
+    """
+    folder_name_path = pathlib.Path(folder_name)
+    coguard_ignore_path = folder_name_path.joinpath(".coguardignore")
+    if not coguard_ignore_path.exists():
+        return []
+    result = []
+    with coguard_ignore_path.open(encoding='utf-8') as ignore_stream:
+        result = ignore_stream.readlines()
+    return [elem for elem in result if elem.strip() and not elem.startswith("#")]
+
+def dry_run_outp(zip_candidate: Tuple[str, Dict]):
+    """
+    The function to output the location of the zip for the
+    dry-run to the user.
+    """
+    if zip_candidate is None:
+        print("No file generated")
+        return
+    zip_location, _ = zip_candidate
+    print("Dry-run complete. You can find the upload-candidate "
+          f"in the following location: {zip_location}")
+
+def upload_and_evaluate_zip_candidate(
+        zip_candidate: Optional[Tuple[str, Dict]],
+        auth_config: Optional[CoGuardCliConfig],
+        deal_type,
+        token: Token,
+        coguard_api_url: str,
+        scan_identifier: str,
+        output_format: str,
+        fail_level: int,
+        organization: Optional[str],
+        ruleset: str):
+    """
+    The common function to upload a zip file, as generated by the
+    helper functions, and evaluate the returned result.
+    """
+    if zip_candidate is None:
+        print(
+            f"{COLOR_YELLOW}Unable to identify any known configuration files.{COLOR_TERMINATION}"
+        )
+        return
+    zip_file, _ = zip_candidate
+    result = api_connection.send_zip_file_for_scanning(
+        zip_file,
+        auth_config.get_username(),
+        token,
+        coguard_api_url,
+        scan_identifier,
+        organization,
+        ruleset
+    )
+    if result is None:
+        print(
+            f"{COLOR_RED} An error occurred while scanning. Please file a "
+            f"bug report, and include the file located at {zip_file}, if possible. "
+            f"{COLOR_TERMINATION}"
+        )
+        sys.exit(1)
+    logging.debug("The result from the api is: %s",
+                  str(result))
+    print(f"{COLOR_CYAN}SCANNING OF{COLOR_TERMINATION} {scan_identifier}"
+          f" {COLOR_CYAN}COMPLETED{COLOR_TERMINATION}")
+    if 'formatted' in output_format:
+        output_result_json_from_coguard(
+            result or {},
+            token,
+            coguard_api_url,
+            auth_config.get_username(),
+            organization
+        )
+    if 'json' in output_format:
+        with pathlib.Path("result.json").open('w', encoding='utf-8') as result_file:
+            json.dump(result or {}, result_file, indent=2)
+            print("JSON file written to `result.json`")
+    if 'markdown' in output_format:
+        translate_result_to_markdown(result, scan_identifier)
+        print("Markdown file written to `result.md`")
+    if 'sarif' in output_format:
+        translate_result_to_sarif(
+            result or {},
+            pathlib.Path('result.sarif.json')
+        )
+        print("Sarif file written to `result.sarif.json`")
+    if deal_type != DealEnum.ENTERPRISE:
+        print("""
+        🔧 Save time. Automatically find and fix vulnerabilities.
+           Upgrade to auto-remediate issues.
+        """)
+    os.remove(zip_file)
+    max_fail_severity = max(
+        entry["rule"]["severity"] for entry in result.get("failed", [])
+    ) if (result and result.get("failed", [])) else 0
+    if max_fail_severity >= fail_level:
+        sys.exit(1)
+
+def output_result_json_from_coguard(
+        result_json: Dict,
+        token: Token,
+        coguard_api_url: str,
+        user_name: Optional[str],
+        organization: Optional[str]):
+    """
+    The function which outputs the result json in a pretty format to the screen.
+
+    :param result_json: The output from the API call to CoGuard.
+    """
+    high_checks = [entry for entry in result_json.get("failed", []) \
+                   if entry["rule"]["severity"] > 3]
+    high_checks.sort(key = lambda x: x["rule"]["severity"], reverse=True)
+    medium_checks = [entry for entry in result_json.get("failed", []) \
+                     if entry["rule"]["severity"] == 3]
+    medium_checks.sort(key = lambda x: x["rule"]["severity"], reverse=True)
+    low_checks = [entry for entry in result_json.get("failed", []) if entry["rule"]["severity"] < 3]
+    low_checks.sort(key = lambda x: x["rule"]["severity"], reverse=True)
+    fixable_check_list = api_connection.get_fixable_rule_list(
+        token,
+        coguard_api_url,
+        user_name,
+        organization
+    )
+    fixable_checks = [entry for entry in result_json.get("failed", []) \
+                      if entry["rule"]["name"] in fixable_check_list]
+    summary = (f'Scan result: {len(result_json.get("failed", []))} checks failed, '
+               f"{COLOR_RED}{len(high_checks)} High{COLOR_TERMINATION}/"
+               f"{COLOR_YELLOW}{len(medium_checks)} Medium{COLOR_TERMINATION}/"
+               f"{COLOR_GRAY}{len(low_checks)} Low{COLOR_TERMINATION} "
+               f"(🔧 {len(fixable_checks)} candidates for auto-remediation)")
+    print(summary)
+    for entry in high_checks:
+        print_failed_check(COLOR_RED, entry, fixable_check_list)
+    for entry in medium_checks:
+        print_failed_check(COLOR_YELLOW, entry, fixable_check_list)
+    for entry in low_checks:
+        print_failed_check(COLOR_GRAY, entry, fixable_check_list)
+    print(summary)
+
+def print_failed_check(color: str,
+                       entry: Dict,
+                       fixable_checks: List[str]):
+    """
+    This is the function to print a failed check entry, given a color.
+
+    :param color: The color to use. see e.g. :const:`COLOR_RED`
+    :param entry: The entry as returned by the CoGuard API.
+    """
+    reference = extract_reference_string(entry)
+    fix_icon = "🔧 " if entry["rule"]["name"] in fixable_checks else ""
+    print(
+        f'{fix_icon}{color} X Severity {entry["rule"]["severity"]}: '
+        f'{entry["rule"]["name"]}{COLOR_TERMINATION}'
+        f'{reference}'
+    )
+    prefix = "Documentation: "
+    try:
+        terminal_size = os.get_terminal_size().columns
+    except OSError:
+        terminal_size = 80
+    wrapper = textwrap.TextWrapper(
+        initial_indent=prefix,
+        width=max(80, terminal_size//2),
+        subsequent_indent=' '*len(prefix)
+    )
+    documentation_candidate = entry["rule"]["documentation"]
+    logging.debug("The candidate for the documentation presentation is: %s",
+                  documentation_candidate)
+    if isinstance(documentation_candidate, str):
+        print(wrapper.fill(entry["rule"]["documentation"]))
+    else:
+        description = documentation_candidate["documentation"]
+        remediation = documentation_candidate["remediation"]
+        sources = ",\n".join(documentation_candidate["sources"])
+        if "scenarios" in documentation_candidate:
+            scenario_string = "\nReferences for the specific ruleset: " + ",\n".join(
+                documentation_candidate["scenarios"]
+            )
+        else:
+            scenario_string = ''
+        documentation_string = f"""
+        {description}
+
+        Remediation: {remediation}
+        {scenario_string}
+
+        Source:
+        {sources}
+        """.replace("        ", "")
+        print(wrapper.fill(documentation_string))
+
+def extract_reference_string(entry_dict: Dict):
+    """
+    This is a helper function to extract the respective file in the manifest
+    corresponding to an entry in the failed rules.
+    """
+    config_file_in_entry = entry_dict.get("config_file", {})
+    from_line = entry_dict.get("fromLine", None)
+    to_line = entry_dict.get("toLine", None)
+    if (from_line is not None and to_line is not None) and (from_line != 0 and to_line != 1):
+        from_to_line = f", {from_line}-{to_line}"
+    else:
+        from_to_line = ""
+    if not config_file_in_entry:
+        return ""
+    reference_path = pathlib.Path(
+        config_file_in_entry.get("subPath", ".")
+    ).joinpath(
+        config_file_in_entry.get("fileName", ".")
+    )
+    return f" (affected files: {reference_path}{from_to_line})"
